Move QualificationsInterfaceController into qualifications folder
Consistent with how we organise Check and Support controllers.
